[#2224] Move exceptions out from workshops.util
