+++ amy/workshops/management/commands/check_for_workshop_websites_updates.py
@@ -11,12 +11,9 @@ from github.GithubException import GithubException
 import requests
 from rest_framework.utils.encoders import JSONEncoder
 
+from workshops.exceptions import WrongWorkshopURL
 from workshops.models import Event
-from workshops.util import (
-    WrongWorkshopURL,
-    fetch_workshop_metadata,
-    parse_workshop_metadata,
-)
+from workshops.util import fetch_workshop_metadata, parse_workshop_metadata
 
 
 def datetime_match(string):
